Finalizacion funcionamiento empleado

+++ database/migrations/2024_11_25_033502_create_empleados_table.php
@@ -15,8 +15,8 @@ return new class extends Migration
             $table->id();
             $table->string("nombres");
             $table->string("Apellidos");
-            $table->integer("identificacion");
-            $table->integer("telefono");
+            $table->bigInteger("identificacion");
+            $table->bigInteger("telefono");
             $table->unsignedBigInteger('id_ciudad');
             $table->unsignedBigInteger('id_departamento');
